Fix Linux build warnings and GCC incompatibilities
Resolve Linux CI job failure by fixing GCC-incompatible client callsites and hardening several Linux-specific warnings and portability issues. Key changes:

- Add docs: nightly recovery log and update SWOT roadmap to reference the fix.
- Meson: separate C-only warning args (avoid -Wmissing-prototypes on C++), introduce engine_c_args and propagate to C targets (renderers, q2proto, executables).
- src/client/effects.cpp: add scalar overload and replace VEC3 callsites to avoid taking addresses of temporary arrays.
- src/client/view.cpp: fix entity loop index type to match r_numentities, add fade_distance overload and update light fade callsite.
- src/client/sound/qal.cpp: use size_t for section/driver loops to silence GCC array-count comparisons.
- src/common/bsp.c: rework patched-PVS path assembly to bounded concatenation with explicit length checks (avoid strncpy/strncat overflow diagnostics).
- src/common/zone.c: guard printing of z_tagnames with a non-null fallback string.
- src/server/mvd/parse.c: remove warning-prone leaf1 lifetime by caching leaf area before multicast checks.

These changes address the GCC/Linux rejection of address-of-temporary-array conversions, remove inappropriate C-only warnings from C++ compilation units, and reduce several source-level warnings that made the Linux build fail in CI.

+++ src/client/sound/qal.cpp
@@ -148,7 +148,7 @@ void QAL_Shutdown(void)
         device = NULL;
     }
 
-    for (int i = 0; i < q_countof(sections); i++) {
+    for (size_t i = 0; i < q_countof(sections); i++) {
         const alsection_t *sec = &sections[i];
         const alfunction_t *func;
 
@@ -205,7 +205,7 @@ int QAL_Init(void)
     const alsection_t *sec;
     const alfunction_t *func;
     ALCint major, minor;
-    int i;
+    size_t i;
 
     al_device = Cvar_Get("al_device", "", 0);
     al_hrtf = Cvar_Get("al_hrtf", "0", 0);
